StoreApp/requirement.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.9
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-decouple==2.1
django-filter==24.2
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-paypal==2.1
django-taggit==5.0.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
idna==3.7
kombu==5.3.4
pillow==10.2.0
prompt-toolkit==3.0.43
PyJWT==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
